refactor(core): make more geoarrow public
This lets use get JSON w/o having to get STAC Items
